pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neuravec-llm"
version = "0.2.0"
description = "Async Azure OpenAI toolkit — structured output, agent pipelines, vision, token tracking"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "neuravec"}]
keywords = ["azure", "openai", "llm", "async", "pydantic", "structured-output"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/neuravec/LLMService"
Repository = "https://github.com/neuravec/LLMService"

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.setuptools.packages.find]
include = ["llm_service*"]
